Move the save dropbox handler out of the routes file
