+++ index.js
@@ -1,6 +1,5 @@
 const {
   getDropbox,
-  saveDropbox,
   getDropboxes,
   createEmptyDropbox,
   deleteDocument,
@@ -11,7 +10,6 @@ const {
   authorize,
   updateArchiveStatus
 } = require('./lib/Dependencies');
-const querystring = require('querystring');
 const api = require('lambda-api')();
 
 const redirectToDropboxesIfAuth = (req, res, next) => {
@@ -160,34 +158,8 @@ api.post(
   }
 );
 
-const saveDropboxHandler = async event => {
-  try {
-    if (event.isBase64Encoded) {
-      event.body = Buffer.from(event.body, 'base64').toString();
-    }
-
-    const session = getSession(event.headers);
-    const { dropboxId: pathDropboxId } = event.pathParameters;
-
-    if (!session || session.dropboxId !== pathDropboxId) {
-      return { statusCode: 404 };
-    }
-
-    const { dropboxId } = session;
-    await saveDropbox(dropboxId, querystring.parse(event.body));
-
-    return {
-      statusCode: 302,
-      headers: { Location: `/dropboxes/${dropboxId}` }
-    };
-  } catch (err) {
-    console.log(err);
-  }
-};
-
 module.exports = {
-  appHandler: async (event, context) => {
+  handler: async (event, context) => {
     return await api.run(event, context);
-  },
-  saveDropboxHandler
+  }
 };
